+++ mirrord-layer/src/socket/hooks.rs
@@ -98,9 +98,7 @@ pub(crate) unsafe extern "C" fn connect_detour(
     match address {
         Ok(address) => {
             let (Ok(result) | Err(result)) = connect(sockfd, address).map_err(|fail| match fail {
-                HookError::LocalFDNotFound(_)
-                | HookError::SocketInvalidState(_)
-                | HookError::BypassedPort(_) => {
+                HookError::LocalFDNotFound(_) | HookError::BypassedPort(_) => {
                     warn!("connect_detour -> bypassed with {:#?}", fail);
                     FN_CONNECT(sockfd, raw_address, address_length)
                 }